video_player: allow passing a second audioSource and videoUrl change on runtime
This video_player now accepts a second audioSource when initializating the
controller which will merge with the current videoSource and play in sync.

Also, now current video stream playing can be changed via controller
function changeVideoUrl(url) which will cause the new stream to merge with
the previously provided audioSource (if provided) and then continue
playing. (This feature is intended for multiple video quality sources. for
changing the source quality)